requirements-dev.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.5
django-allauth==0.44.0
django-cors-headers==3.6.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
progress==1.5
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.6
python-decouple==3.4
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
webencodings==0.5.1
